+++ Plutarch/Either.hs
@@ -36,7 +36,6 @@ import Generics.SOP qualified as SOP
 import Plutarch.Builtin.Bool (
   PBool (PFalse, PTrue),
   pif,
-  pif',
  )
 import Plutarch.Builtin.Data (
   PAsData,
@@ -209,19 +208,19 @@ instance
   {-# INLINEABLE pmax #-}
   pmax t1 t2 = pmatch t1 $ \case
     PDLeft t1' -> pmatch t2 $ \case
-      PDLeft t2' -> pif' # (pfromData t1' #< pfromData t2') # t2 # t1
+      PDLeft t2' -> pif (pfromData t1' #< pfromData t2') t2 t1
       PDRight _ -> t2
     PDRight t1' -> pmatch t2 $ \case
       PDLeft _ -> t1
-      PDRight t2' -> pif' # (pfromData t1' #< pfromData t2') # t2 # t1
+      PDRight t2' -> pif (pfromData t1' #< pfromData t2') t2 t1
   {-# INLINEABLE pmin #-}
   pmin t1 t2 = pmatch t1 $ \case
     PDLeft t1' -> pmatch t2 $ \case
-      PDLeft t2' -> pif' # (pfromData t1' #< pfromData t2') # t1 # t2
+      PDLeft t2' -> pif (pfromData t1' #< pfromData t2') t1 t2
       PDRight _ -> t1
     PDRight t1' -> pmatch t2 $ \case
       PDLeft _ -> t2
-      PDRight t2' -> pif' # (pfromData t1' #< pfromData t2') # t1 # t2
+      PDRight t2' -> pif (pfromData t1' #< pfromData t2') t1 t2
 
 -- | @since 1.10.0
 instance PlutusType (PEitherData a b) where